Smarter state detection to save time.

+++ fogAutomatedTesting/captureImage.sh
@@ -45,12 +45,19 @@ chown -R $sharePermissions $shareDir
 
 echo "$(date +%x_%r) Restoring snapshot \"$snapshot\" to \"$vmGuest\"" >> $output
 ssh -o ConnectTimeout=$sshTimeout $hostsystem "virsh snapshot-revert $vmGuest $snapshot" > /dev/null 2>&1
+
+
+
 #Gracefully shutdown VM incase it's on.
-echo "$(date +%x_%r) Asking \"$vmGuest\" to gracefully shutdown." >> $output
-ssh -o ConnectTimeout=$sshTimeout $hostsystem "virsh shutdown \"$vmGuest\" > /dev/null 2>&1
-sleep 30
-#Kill it if it lags.
-ssh -o ConnectTimeout=$sshTimeout $hostsystem "virsh destroy \"$vmGuest\" > /dev/null 2>&1
+if [[ $(virsh domstate $vmGuest) == "running" ]]; then
+    echo "$(date +%x_%r) Asking \"$vmGuest\" to gracefully shutdown." >> $output
+    ssh -o ConnectTimeout=$sshTimeout $hostsystem "virsh shutdown \"$vmGuest\" > /dev/null 2>&1
+    sleep 30
+    ssh -o ConnectTimeout=$sshTimeout $hostsystem "virsh destroy \"$vmGuest\" > /dev/null 2>&1
+else
+    echo "$(date +%x_%r) \"$vmGuest\" is already shutdown." >> $output
+fi
+
 
 echo "$(date +%x_%r) Queuing the capture job on the server." >> $output
